chore: added auth directory and moved login to auth.
